rename dots_utilities to libdots
