+++ scripts/compile_protobuf.sh
@@ -1,2 +1,2 @@
 #!/bin/sh
-python3 -m grpc_tools.protoc -I=src/dots_utilities/io/message_definitions --python_out=src/dots_utilities/io/messages --pyi_out=./src/dots_utilities/io/messages  src/dots_utilities/io/message_definitions/*.proto
+python3 -m grpc_tools.protoc -I=src/libdots/io/message_definitions --python_out=src/libdots/io/messages --pyi_out=./src/libdots/io/messages  src/libdots/io/message_definitions/*.proto
