+++ src/libdots/io/input_data_inventory.py
