+++ src/libdots/io/message_definitions/healthcheck.proto
