+++ src/libdots/io/message_definitions/lifecycle.proto
